+++ include/operon/core/variable.hpp
@@ -10,7 +10,7 @@ namespace Operon {
 struct Variable {
     std::string Name;
     Operon::Hash Hash{0};
-    size_t Index = size_t{0};
+    int64_t Index = size_t{0};
 
     constexpr auto operator==(Variable const& rhs) const noexcept -> bool {
         return std::tie(Name, Hash, Index) == std::tie(rhs.Name, rhs.Hash, rhs.Index);
